Fine module fixes, nav/layout updates, maintenance Pay Bill operator session
- Fine: summary grouping, model types, UI alignment; layout title

- Home/Maintenance/EBill sidebars: Audit, Settings, Tariff, links reorder

- Login + MaintenanceBill PaymentForm: resolve OperatorId; remove EF6 AsNoTracking conflict

Made-with: Cursor

+++ BMSBT/Models/Fine.cs
@@ -47,16 +47,16 @@ namespace BMSBT.Models
 
 
         [Required(ErrorMessage = "Fine Amount is required")]
-        public decimal? FineAmount { get; set; }
+        public int? FineAmount { get; set; }
 
         [Required(ErrorMessage = "Waived Amount is required")]
-        public decimal? WaivedAmount { get; set; }
+        public int? WaivedAmount { get; set; }
         
         [Required(ErrorMessage = "Adjustment Amount is required")]
-        public decimal? AdjustmentAmount { get; set; }
+        public int? AdjustmentAmount { get; set; }
 
         [Required(ErrorMessage = "Fine To Charge is required")]
-        public decimal FineToCharge { get; set; }
+        public int FineToCharge { get; set; }
 
         public string? Comments { get; set; }
         public string? FineEnteredBy { get; set; }
